Use existing function to call photon mosaic
