+++ tests/test_integration/test_whole_workflow.py
@@ -8,6 +8,7 @@ import yaml
 from tifffile import imread
 
 from photon_mosaic import get_snakefile_path
+from tests.conftest import run_photon_mosaic
 
 
 def run_snakemake(workdir, configfile, dry_run=False):
@@ -246,19 +247,9 @@ def test_snakemake_with_contrast(snake_test_env, test_config_with_contrast):
 
 def test_photon_mosaic_cli_dry_run(snake_test_env):
     """Test that photon-mosaic can do a dry run."""
-    cmd = [
-        "photon-mosaic",
-        "--config",
-        str(snake_test_env["configfile"]),
-    ]
-
-    result = subprocess.run(
-        cmd,
-        cwd=snake_test_env["workdir"],
-        capture_output=True,
-        text=True,
-        encoding="utf-8",
-        errors="replace",
+    result = run_photon_mosaic(
+        snake_test_env["workdir"],
+        snake_test_env["configfile"],
     )
 
     assert result.returncode == 0, (
@@ -269,19 +260,9 @@ def test_photon_mosaic_cli_dry_run(snake_test_env):
 
 def test_photon_mosaic_cli(snake_test_env):
     """Test photon-mosaic pipeline."""
-    cmd = [
-        "photon-mosaic",
-        "--config",
-        str(snake_test_env["configfile"]),
-    ]
-
-    result = subprocess.run(
-        cmd,
-        cwd=snake_test_env["workdir"],
-        capture_output=True,
-        text=True,
-        encoding="utf-8",
-        errors="replace",
+    result = run_photon_mosaic(
+        snake_test_env["workdir"],
+        snake_test_env["configfile"],
     )
 
     assert result.returncode == 0, (
